backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
redis==5.0.1
openai==1.3.0
APScheduler==3.10.4
psycopg2-binary==2.9.9
